requirements.txt
alembic==0.9.9
aniso8601==3.0.0
astroid==1.6.3
attrs==17.4.0
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
decorator==4.3.0
dnspython==1.15.0
docopt==0.6.2
Flask==0.12.2
Flask-Cors==3.0.6
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
pep8==1.7.1
pluggy==0.6.0
psycopg2-binary==2.7.4
py==1.5.3
pycodestyle==2.4.0
pylint==1.8.4
pytest==3.5.1
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.4
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.7
urllib3==1.22
validators==0.12.1
Werkzeug==0.14.1
wrapt==1.10.11
